rename Result component and path in Report
