+++ frontend/src/components/Report.vue
